extract printWord
