+++ src/language-markdown/print/word.js
@@ -1,0 +1,56 @@
+import { PUNCTUATION_REGEXP } from "../constants.evaluate.js";
+import { isAutolink } from "../utils.js";
+
+/**
+ * @import AstPath from "../../common/ast-path.js"
+ * @import {Doc} from "../../document/index.js"
+ */
+
+/**
+ * @param {AstPath} path
+ * @return {Doc}
+ */
+function printWord(path) {
+  const { node } = path;
+  let escapedValue = node.value
+    .replaceAll("*", String.raw`\*`) // escape all `*`
+    .replaceAll(
+      new RegExp(
+        [
+          `(^|${PUNCTUATION_REGEXP.source})(_+)`,
+          `(_+)(${PUNCTUATION_REGEXP.source}|$)`,
+        ].join("|"),
+        "gu",
+      ),
+      (_, text1, underscore1, underscore2, text2) =>
+        (underscore1
+          ? `${text1}${underscore1}`
+          : `${underscore2}${text2}`
+        ).replaceAll("_", String.raw`\_`),
+    ); // escape all `_` except concating with non-punctuation, e.g. `1_2_3` is not considered emphasis
+
+  const isFirstSentence = (node, name, index) =>
+    node.type === "sentence" && index === 0;
+  const isLastChildAutolink = (node, name, index) =>
+    isAutolink(node.children[index - 1]);
+
+  if (
+    escapedValue !== node.value &&
+    (path.match(undefined, isFirstSentence, isLastChildAutolink) ||
+      path.match(
+        undefined,
+        isFirstSentence,
+        (node, name, index) => node.type === "emphasis" && index === 0,
+        isLastChildAutolink,
+      ))
+  ) {
+    // backslash is parsed as part of autolinks, so we need to remove it
+    escapedValue = escapedValue.replace(/^(\\?[*_])+/u, (prefix) =>
+      prefix.replaceAll("\\", ""),
+    );
+  }
+
+  return escapedValue;
+}
+
+export { printWord };
